Fixed thread exit hanging (hack until fix by Godot)

+++ scripts/autoloads/GDNShell.gd
@@ -23,13 +23,13 @@ func GDN_INIT():
 	GDNSHELL.connect("child_process_started", self, "_child_process_started")
 	GDNSHELL.connect("child_process_stopped", self, "_child_process_stopped")
 
-var is_stopping = false
-func thread_subroutine(cmd):
+func _THREAD_SUBROUTINE(cmd):
 	# this will internally loop and update, function will resume on close
 	Log.generic(self, str(":: Entering thread ", shell_thread.get_id()))
 	GDNSHELL.spawn(cmd)
 
 	# cleanup thread
+	shell_thread.call_deferred("wait_to_finish")
 	Log.generic(self, str(":: Exiting thread ", shell_thread.get_id()))
 
 var shell_cmd = "E:/Git/CppTestApp/cmake-build-debug/CppTestApp.exe"
@@ -40,26 +40,22 @@ func start():
 	stop()
 
 	# clear terminal & start thread
-	clear()
-	shell_thread.start(self, "thread_subroutine", shell_cmd)
+	terminal_node.text = ""
+	shell_thread.start(self, "_THREAD_SUBROUTINE", shell_cmd)
 func stop():
 	assert(GDNSHELL != null)
-	if is_stopping:
-		return
-	is_stopping = true
-	if GDNSHELL != null && shell_thread != null && shell_thread.is_active():
+	if GDNSHELL != null && shell_thread != null && shell_thread.is_active() && shell_thread.is_alive():
 		Log.generic(self, str("Killing child process on thread ", shell_thread.get_id()))
 		GDNSHELL.kill()
 		if shell_thread.is_alive():
 			shell_thread.wait_to_finish()
 	else:
 		Log.generic(self, "No process is running!")
-	is_stopping = false
-
 
 func sync_text(line, size):
 	if GDNSHELL != null:
-		var buffer = GDNSHELL.fetch_at_line(line, size)
+#		var buffer = GDNSHELL.fetch_at_line(line, size)
+		var buffer = GDNSHELL.fetch_since_last_time()
 		return buffer
 	return null
 
@@ -70,11 +66,10 @@ func send_string(s : String):
 		else:
 			print("failure...")
 func clear():
-	return
 	assert(terminal_node != null)
 	assert(GDNSHELL != null)
-	terminal_node.clear()
 	GDNSHELL.clear()
+	terminal_node.text = ""
 
 func _exit_tree():
 	# close any open session when shutting down
